requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
backoff==2.2.1
    # via opentelemetry-exporter-otlp-proto-grpc
build==1.0.3
    # via pip-tools
click==8.1.7
    # via pip-tools
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
exceptiongroup==1.1.3
    # via pytest
googleapis-common-protos==1.60.0
    # via opentelemetry-exporter-otlp-proto-grpc
grpcio==1.59.0
    # via opentelemetry-exporter-otlp-proto-grpc
importlib-metadata==6.8.0
    # via
    #   build
    #   opentelemetry-api
iniconfig==2.0.0
    # via pytest
opentelemetry-api==1.19.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-celery
    #   opentelemetry-sdk
opentelemetry-exporter-otlp-proto-common==1.19.0
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-grpc==1.19.0
    # via -r requirements.in
opentelemetry-instrumentation==0.40b0
    # via opentelemetry-instrumentation-celery
opentelemetry-instrumentation-celery==0.40b0
    # via -r requirements.in
opentelemetry-proto==1.19.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.19.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-semantic-conventions==0.40b0
    # via
    #   opentelemetry-instrumentation-celery
    #   opentelemetry-sdk
packaging==23.2
    # via
    #   build
    #   pytest
pip-tools==6.13.0
    # via -r requirements.in
pluggy==1.3.0
    # via pytest
protobuf==4.24.4
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
pyproject-hooks==1.0.0
    # via build
pytest==7.4.2
    # via -r requirements.in
tomli==2.0.1
    # via
    #   build
    #   pyproject-hooks
    #   pytest
typing-extensions==4.8.0
    # via opentelemetry-sdk
wheel==0.41.2
    # via pip-tools
wrapt==1.15.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
